backend/requirements.txt
fastapi==0.111.0
uvicorn==0.40.0
pydantic==2.12.5
PyJWT==2.11.0
email-validator==2.3.0
requests==2.31.0
beautifulsoup4==4.12.3
